crio, deploy: provide manifests to deploy controller for this runtime
This PR changes the manifest templating to also provide a manifest for
CRIO based installations.

This manifest features a different runtime configuration, which will
cause the controller to use a different container runtime to access the
pod's sandbox ID, and namespace.

The golang code is also updated to use this new runtime, when
configured.

Signed-off-by: Miguel Duarte Barroso <mdbarroso@redhat.com>