e2e tests: align the CNI plugin versions on NADs
We plan on shiping this prohect via the CNAO operator - [0]. It runs the
e2e tests of every project it ships as part of its PR verification
strategy. For now, the version of the plugins it ships does not allow for
for CNI plugins version 1.0.0. As such, we will for now fall-back to this
old version.

Once that is updated, we will revert this commit.

[0] - https://github.com/kubevirt/cluster-network-addons-operator

Signed-off-by: Miguel Duarte Barroso <mdbarroso@redhat.com>