e2e tests: add test asserting reconcile loop
Add an e2e test that attempts to add an interface to a running pod;
since originally one of the pod's network selection elements does not
feature the interface name request, the dynamic networks controller
cannot reconcile it.

We then update the pod's network selection elements to feature the
correct interface name, which causes the reconcile loop to succeed, and
the previously added interface to be added.

Signed-off-by: Miguel Duarte Barroso <mdbarroso@redhat.com>