ci, build: use the correct dockerfile image
We must use the dockerfile for the thick plugin version.

Up to now this was not an issue since the thin version featured the `cp`
command, which is used on the thick-plugin initcontainer. Now that the
thin plugin version daemonset spec was updated (it no longer features
the `cp` command), the thick plugin version init container fails to
perform its task.

Signed-off-by: Miguel Duarte Barroso <mdbarroso@redhat.com>