fix: Replace j2cli with jinjanator in GitHub workflows
j2cli is unmaintained and no longer works with python3.12. It is
therefore replaced with an active fork called jinjanator.

The improper use of the '-e' flag on the j2 binary is fixed. Since the
env is used as data source there is also no need to import variables
explicitly again.

Signed-off-by: Felix Matouschek <fmatouschek@redhat.com>