Fix network status on failing attachement add/delete
When adding 2 attachements and the second was failing, the status was
not reported in the annotation. Same behavior was happening for
deletion.
This is now fixed, and events are also reported in case of failure on
adding/deleting annotations.
The order in which attachements are handled is now deterministic to
fulfill some special use cases (e.g. macvlan attachement + vlan
attachement based on the macvlan) and to improve testing and
troubleshooting.
This behavior should be now locked with the pod controller unit tests.